Add files via upload
this my first attempt at coding a very simple and basic program that highlights the part of a Fundus fluorescein angiography video which has equally prominent arterial and venous components by looking at the overall brightness of each frame of a video and selecting the frame with the maximum brightness. I had previously helped come up with a similar software with my cousin by contributing from the capacity of a medical person. But since I have started to learn coding as a hobby I thought I would try and give it a shot on doing it all by myself.